+++ crud/borders.lua
@@ -1,6 +1,5 @@
 local checks = require('checks')
 local errors = require('errors')
-local vshard = require('vshard')
 
 local const = require('crud.common.const')
 local dev_checks = require('crud.common.dev_checks')
@@ -71,9 +70,8 @@ local function call_get_border_on_router(vshard_router, border_name, space_name,
     checks('table', 'string', 'string', '?string|number', {
         timeout = '?number',
         fields = '?table',
+        vshard_router = '?string|table',
     })
-
-    opts = opts or {}
 
     local replicasets = vshard_router:routeall()
     local space = utils.get_space(space_name, replicasets)
@@ -160,7 +158,11 @@ local function call_get_border_on_router(vshard_router, border_name, space_name,
 end
 
 local function get_border(border_name, space_name, index_name, opts)
-    local vshard_router = vshard.router.static
+    opts = opts or {}
+    local vshard_router, err = utils.get_vshard_router_instance(opts.vshard_router)
+    if err ~= nil then
+        return nil, BorderError:new(err)
+    end
 
     return schema.wrap_func_reload(vshard_router, call_get_border_on_router,
         border_name, space_name, index_name, opts
@@ -182,6 +184,11 @@ end
 --
 -- @tparam ?table opts.fields
 --  Field names for getting only a subset of fields
+--
+-- @tparam ?string|table opts.vshard_router
+--  Cartridge vshard group name or vshard router instance.
+--  Set this parameter if your space is not a part of the
+--  default vshard cluster.
 --
 -- @return[1] result
 -- @treturn[2] nil
@@ -205,6 +212,11 @@ end
 --
 -- @tparam ?table opts.fields
 --  Field names for getting only a subset of fields
+--
+-- @tparam ?string|table opts.vshard_router
+--  Cartridge vshard group name or vshard router instance.
+--  Set this parameter if your space is not a part of the
+--  default vshard cluster.
 --
 -- @return[1] result
 -- @treturn[2] nil