+++ crud/count.lua
@@ -1,6 +1,5 @@
 local checks = require('checks')
 local errors = require('errors')
-local vshard = require('vshard')
 local fiber = require('fiber')
 
 local call = require('crud.common.call')
@@ -120,9 +119,8 @@ local function call_count_on_router(vshard_router, space_name, user_conditions, 
         prefer_replica = '?boolean',
         balance = '?boolean',
         mode = '?string',
+        vshard_router = '?string|table',
     })
-
-    opts = opts or {}
 
     if opts.yield_every ~= nil and opts.yield_every < 1 then
         return nil, CountError:new("yield_every should be > 0")
@@ -308,6 +306,11 @@ end
 -- @tparam ?string opts.mode
 --  vshard call mode, default value is `read`
 --
+-- @tparam ?string|table opts.vshard_router
+--  Cartridge vshard group name or vshard router instance.
+--  Set this parameter if your space is not a part of the
+--  default vshard cluster.
+--
 -- @return[1] number
 -- @treturn[2] nil
 -- @treturn[2] table Error description
@@ -322,9 +325,15 @@ function count.call(space_name, user_conditions, opts)
         prefer_replica = '?boolean',
         balance = '?boolean',
         mode = '?string',
+        vshard_router = '?string|table',
     })
 
-    local vshard_router = vshard.router.static
+    opts = opts or {}
+
+    local vshard_router, err = utils.get_vshard_router_instance(opts.vshard_router)
+    if err ~= nil then
+        return nil, CountError:new(err)
+    end
 
     return schema.wrap_func_reload(vshard_router, sharding.wrap_method, call_count_on_router,
                                    space_name, user_conditions, opts)