+++ crud/replace.lua
@@ -1,6 +1,5 @@
 local checks = require('checks')
 local errors = require('errors')
-local vshard = require('vshard')
 
 local call = require('crud.common.call')
 local const = require('crud.common.const')
@@ -62,9 +61,8 @@ local function call_replace_on_router(vshard_router, space_name, original_tuple,
         bucket_id = '?number|cdata',
         add_space_schema_hash = '?boolean',
         fields = '?table',
+        vshard_router = '?string|table',
     })
-
-    opts = opts or {}
 
     local space, err = utils.get_space(space_name, vshard_router:routeall())
     if err ~= nil then
@@ -142,6 +140,11 @@ end
 --  Bucket ID
 --  (by default, it's vshard.router.bucket_id_strcrc32 of primary key)
 --
+-- @tparam ?string|table opts.vshard_router
+--  Cartridge vshard group name or vshard router instance.
+--  Set this parameter if your space is not a part of the
+--  default vshard cluster.
+--
 -- @return[1] object
 -- @treturn[2] nil
 -- @treturn[2] table Error description
@@ -152,9 +155,15 @@ function replace.tuple(space_name, tuple, opts)
         bucket_id = '?number|cdata',
         add_space_schema_hash = '?boolean',
         fields = '?table',
+        vshard_router = '?string|table',
     })
 
-    local vshard_router = vshard.router.static
+    opts = opts or {}
+
+    local vshard_router, err = utils.get_vshard_router_instance(opts.vshard_router)
+    if err ~= nil then
+        return nil, ReplaceError:new(err)
+    end
 
     return schema.wrap_func_reload(vshard_router, sharding.wrap_method, call_replace_on_router,
                                    space_name, tuple, opts)
@@ -183,9 +192,15 @@ function replace.object(space_name, obj, opts)
         bucket_id = '?number|cdata',
         add_space_schema_hash = '?boolean',
         fields = '?table',
+        vshard_router = '?string|table',
     })
 
-    local vshard_router = vshard.router.static
+    opts = opts or {}
+
+    local vshard_router, err = utils.get_vshard_router_instance(opts.vshard_router)
+    if err ~= nil then
+        return nil, ReplaceError:new(err)
+    end
 
     -- replace can fail if router uses outdated schema to flatten object
     opts = utils.merge_options(opts, {add_space_schema_hash = true})