+++ crud/select/compat/select.lua
@@ -1,6 +1,5 @@
 local checks = require('checks')
 local errors = require('errors')
-local vshard = require('vshard')
 
 local const = require('crud.common.const')
 local utils = require('crud.common.utils')
@@ -200,6 +199,8 @@ function select_module.pairs(space_name, user_conditions, opts)
         prefer_replica = '?boolean',
         balance = '?boolean',
         timeout = '?number',
+
+        vshard_router = '?string|table',
     })
 
     opts = opts or {}
@@ -208,7 +209,10 @@ function select_module.pairs(space_name, user_conditions, opts)
         error(string.format("Negative first isn't allowed for pairs"))
     end
 
-    local vshard_router = vshard.router.static
+    local vshard_router, err = utils.get_vshard_router_instance(opts.vshard_router)
+    if err ~= nil then
+        error(err)
+    end
 
     local iterator_opts = {
         after = opts.after,
@@ -266,9 +270,9 @@ local function select_module_call_xc(vshard_router, space_name, user_conditions,
         prefer_replica = '?boolean',
         balance = '?boolean',
         timeout = '?number',
-    })
 
-    opts = opts or {}
+        vshard_router = '?string|table',
+    })
 
     if opts.first ~= nil and opts.first < 0 then
         if opts.after == nil then
@@ -324,7 +328,12 @@ local function select_module_call_xc(vshard_router, space_name, user_conditions,
 end
 
 function select_module.call(space_name, user_conditions, opts)
-    local vshard_router = vshard.router.static
+    opts = opts or {}
+
+    local vshard_router, err = utils.get_vshard_router_instance(opts.vshard_router)
+    if err ~= nil then
+        return nil, SelectError:new(err)
+    end
 
     return SelectError:pcall(sharding.wrap_select_method, vshard_router, select_module_call_xc,
                              space_name, user_conditions, opts)