+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+* Rework `NonInitialized` error message to be more helpful for
+  troubleshooting (#326).
+
 ## [0.14.0] - 13-09-22
 
 ### Added